+++ CHANGELOG.md
@@ -59,6 +59,7 @@ To learn more about active deprecations, we recommend checking [GitHub Discussio
 ### New
 
 - TODO ([#XXX](https://github.com/kedacore/keda/issues/XXX))
+- **General**: Declarative scaler config parsing ([#5037](https://github.com/kedacore/keda/issues/5037))
 
 #### Experimental